refactor: create chainable transformers
