+++ transforms/message-remove-rounded-corners/message-remove-rounded-corners.js
@@ -1,8 +1,7 @@
 /*
  * Convert all <Message as="" /> <Message variant="" />
  */
-import removeAttribute from "../removeAttribute";
-module.exports = removeAttribute(
-  "carbon-react/lib/components/message",
-  "roundedCorners"
+import { run, removeAttribute } from "../builder";
+module.exports = run(
+  removeAttribute("carbon-react/lib/components/message", "roundedCorners")
 );
