--- transforms/removeAttribute.js
@@ -1,3 +1,0 @@
-import renameAttribute from "./renameAttribute";
-
-module.exports = (path, attribute) => renameAttribute(path, attribute);
